requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.1
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
